fix: Disable the func for parsing individual IP range string
In the past, Parse() should be used to parse a single IP range string,
while ParseRanges() used to parse multiple IP range strings. I think
this approach is not very appropriate because they are not uniform and
can cause additional work. Currently, there is only ParseRanges() has
been exposed to public (and it has been renamed Parse()).

And fix a bug when comparing a 16-bit format IPv4 address with a 4-bit
format IPv4 address.

Signed-off-by: iiiceoo <iiiceoo@foxmail.com>